refactor: move initializeAccountTree to authStateMachine
