+++ app/components/UI/EvmAccountSelectorList/EvmAccountSelectorList.types.ts
@@ -65,3 +65,8 @@ export interface EvmAccountSelectorListProps
    */
   privacyMode?: boolean;
 }
+
+export interface AccountSection {
+  title: string;
+  data: Account[];
+}
