+++ app/core/Authentication/Authentication.ts
@@ -39,14 +39,6 @@ import {
   WalletClientType,
 } from '../SnapKeyring/MultichainWalletSnapClient';
 ///: END:ONLY_INCLUDE_IF
-import {
-  assertMultichainAccountsFeatureFlagType,
-  isMultichainAccountsFeatureEnabled,
-  MULTI_CHAIN_ACCOUNTS_FEATURE_VERSION_1,
-  MULTI_CHAIN_ACCOUNTS_FEATURE_VERSION_2,
-  MultichainAccountsFeatureFlag,
-} from '../../selectors/featureFlagController/multichainAccounts/enabledMultichainAccounts';
-import { Json } from '@metamask/utils';
 
 /**
  * Holds auth data used to determine auth configuration
@@ -231,34 +223,6 @@ class AuthenticationService {
       currentAuthType: AUTHENTICATION_TYPE.PASSWORD,
       availableBiometryType,
     };
-  };
-
-  /**
-   * Checks if multichain accounts are enabled for state 1.
-   * @param remoteFeatureFlags - The remote feature flags object containing the multichain accounts feature flag.
-   * @returns True if the multichain accounts feature is enabled for state 1, false otherwise.
-   */
-  private isMultichainAccountsEnabledForState1 = (remoteFeatureFlags: Json & MultichainAccountsFeatureFlag) => (
-    [MULTI_CHAIN_ACCOUNTS_FEATURE_VERSION_1, MULTI_CHAIN_ACCOUNTS_FEATURE_VERSION_2].some((featureVersion) =>
-      isMultichainAccountsFeatureEnabled(remoteFeatureFlags, featureVersion)
-    )
-  );
-
-  /**
-   * Initializes the account tree and updates the accounts if multichain accounts is enabled
-   */
-  private initializeAccountTree = async (): Promise<void> => {
-    const { AccountTreeController, AccountsController, RemoteFeatureFlagController } = Engine.context;
-    const { enableMultichainAccounts } = RemoteFeatureFlagController.state.remoteFeatureFlags;
-    if(!assertMultichainAccountsFeatureFlagType(enableMultichainAccounts)) {
-      return;
-    }
-    const isMultichainAccountsEnabled = this.isMultichainAccountsEnabledForState1(enableMultichainAccounts);
-
-    if (isMultichainAccountsEnabled) {
-      AccountTreeController.init();
-      await AccountsController.updateAccounts();
-    }
   };
 
   /**
@@ -483,8 +447,6 @@ class AuthenticationService {
       this.retrySolanaDiscoveryIfPending();
       ///: END:ONLY_INCLUDE_IF
 
-      //TODO: Move this logic to the Engine when the account tree state will be persisted
-      this.initializeAccountTree();
       // TODO: Replace "any" with type
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
     } catch (e: any) {
@@ -539,8 +501,6 @@ class AuthenticationService {
       this.retrySolanaDiscoveryIfPending();
       ///: END:ONLY_INCLUDE_IF
 
-      //TODO: Move this logic to the Engine when the account tree state will be persisted
-      this.initializeAccountTree();
       // TODO: Replace "any" with type
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
     } catch (e: any) {
