--- app/multichain-accounts/controllers/account-tree-controller/constants.ts
@@ -1,2 +1,0 @@
-// Default AccountTreeController state
-export const defaultAccountTreeControllerState = {};
