+++ app/store/sagas/index.ts
@@ -20,6 +20,7 @@ import {
 } from './xmlHttpRequestOverride';
 import EngineService from '../../core/EngineService';
 import { AppStateEventProcessor } from '../../core/AppStateEventListener';
+import AccountTreeInitService from '../../multichain-accounts/AccountTreeInitService';
 
 export function* appLockStateMachine() {
   let biometricsListenerTask: Task<void> | undefined;
@@ -50,6 +51,8 @@ export function* authStateMachine() {
     yield take(UserActionType.LOGIN);
     const appLockStateMachineTask: Task<void> = yield fork(appLockStateMachine);
     LockManagerService.startListening();
+    //TODO: Move this logic to the Engine when the account tree state will be persisted
+    AccountTreeInitService.initializeAccountTree();
     // Listen to app lock behavior.
     yield take(UserActionType.LOGOUT);
     LockManagerService.stopListening();
